+++ frontend/src/components/landing/LandingCareers.jsx
@@ -1,0 +1,8 @@
+export default function LandingCareers() {
+    return (
+        <div className="container">
+            <h1>Careers</h1>
+            <p className="lead mt-4">We currently do not have any open positions, but will be posting some soon!</p>
+        </div>
+    )
+}