+++ frontend/src/components/landing/LandingContact.jsx
@@ -1,0 +1,8 @@
+export default function LandingContact() {
+    return (
+        <div className="container">
+            <h1>Contact Us</h1>
+            <p className="lead mt-4">We currently do not have an email set up yet, but will have one set up soon!</p>
+        </div>
+    )
+};