+++ frontend/src/components/landing/LandingFAQs.jsx
@@ -1,0 +1,9 @@
+export default function LandingFAQs() {
+    return (
+        <div className="container">
+            <h1>Frequently Asked Questions (FAQ)</h1>
+            <p className="lead mt-4">This website is still a WIP, and does not have any FAQs as of now.</p>
+            <p className="lead mt-4">Please check back later for updates!</p>
+        </div>
+    )
+}